(SERVER-893) Adopt pl-clojure-style
This commit adds cljfmt as a valid lein alias which will defer to the
plugin lein-cljfmt using the pl-clojure-style repo's styles.

We use a git submodule approach to pulling in pl-clojure-style to ease
the ability in the future of using the tool within an automated process
(e.g. travis-ci) and because of significant use of submodules in other
puppet labs clj projects.