(TK-331) Consume all messages on lifecycle channels on shutdown
This commit adds some code to read all of the remaining messages
off of the core.async lifecycle channels after they are closed.
This ensures that if there are any threads waiting on a blocking
put operation, they'll be freed up.

We also add a test that calls `stop` twice to validate that the
second stop does not execute but also does not hang.