(TK-331) Close lifecycle channels earlier, add log message
This commit addresses some PR feedback.  In particular, we close
the lifecycle channels a little earlier when we are performing
a shutdown, to minimize the possibility that other messages are
being queued on the channels while we're shutting down.

We also add a log message that would explain what is happening
in the (hopefully rare/impossible) case where two shutdowns were
triggered at the same time.