(TK-331) Improve logging / try-catch logic in async loop
This commit re-organizes the try/catch logic inside of the
core.async lifecycle worker loop, because in a previous commit,
we moved the shutdown logic into the core.async loop, meaning
that any exception that was thrown from inside of the loop which
was relying on the ability to deliver the 'shutdown promise' in
order to trigger a subsequent shutdown would no longer work
properly.

We now do the try/catch inside of the loop, so that the result
of delivering a shutdown promise will be that the loop keeps
executing, and thus has an opportunity to process the resulting
shutdown message.

This commit also adds some additional logging to the code where
we drain the lifecycle channels so that there will be some
record of the ignored requests / events.